refactor(api): split configured provider instances
