+++ apps/api/src/providers/instances/admin-bot.ts
@@ -1,0 +1,9 @@
+import { config } from '@rctf/config'
+import { adminBotProviders } from '../admin-bot'
+import { loadProvider } from './load'
+
+export const adminBotProvider = loadProvider(
+  adminBotProviders,
+  config.adminBot?.provider
+)
+export const adminBotEnabled = adminBotProvider !== undefined
