+++ apps/api/src/providers/instances/index.ts
@@ -1,0 +1,8 @@
+import './emails'
+import './uploads'
+import './instancer'
+import './captcha'
+import './messages'
+import './moderation'
+import './admin-bot'
+import './analytics'
