+++ apps/api/src/providers/instances/messages.ts
@@ -1,0 +1,7 @@
+import { config } from '@rctf/config'
+import { messagesProviders } from '../messages'
+import { loadProvider } from './load'
+
+export const bloodBotProviders = config.bloodBot?.destinations.map(
+  ({ provider }) => loadProvider(messagesProviders, provider)!
+)
