+++ apps/api/src/providers/instances/uploads.ts
@@ -1,0 +1,8 @@
+import { config } from '@rctf/config'
+import { uploadProviders } from '../uploads'
+import { loadProvider } from './load'
+
+export const uploadProvider = loadProvider(
+  uploadProviders,
+  config.uploadProvider
+)!
