+++ apps/api/src/routes/v2/integrations/routes/get-admin-bot-job-status.ts
@@ -1,5 +1,5 @@
 import { AdminBotJobStatus, GetAdminBotJobStatusRouteV2 } from '@rctf/types'
-import { adminBotProvider } from '../../../../providers'
+import { adminBotProvider } from '../../../../providers/instances/admin-bot'
 import { getLatestJob } from '../../../../services/admin-bot-jobs'
 import { getChallenge } from '../../../../services/challenges'
 import integrationsGroup from '../group'
